refactor: moved exceptions linked to post in right directory, delete unused exceptions, correct postnotfound and postinvalid handler function
