+++ src/main/java/techquack/com/onestar3gram/exceptions/post/InvalidAltException.java
@@ -1,4 +1,4 @@
-package techquack.com.onestar3gram.exceptions;
+package techquack.com.onestar3gram.exceptions.post;
 
 public class InvalidAltException extends Exception{
     public InvalidAltException(String message) {
