+++ src/main/java/techquack/com/onestar3gram/exceptions/post/InvalidDescriptionException.java
@@ -1,4 +1,4 @@
-package techquack.com.onestar3gram.exceptions;
+package techquack.com.onestar3gram.exceptions.post;
 
 public class InvalidDescriptionException extends Exception{
     public InvalidDescriptionException(String message) {
