+++ src/main/java/techquack/com/onestar3gram/exceptions/post/NegativeLikeNumberException.java
@@ -1,4 +1,4 @@
-package techquack.com.onestar3gram.exceptions;
+package techquack.com.onestar3gram.exceptions.post;
 
 public class NegativeLikeNumberException extends  Exception {
     public NegativeLikeNumberException(String message) {
