+++ src/main/java/techquack/com/onestar3gram/exceptions/post/PostNotFoundException.java
@@ -1,4 +1,4 @@
-package techquack.com.onestar3gram.exceptions;
+package techquack.com.onestar3gram.exceptions.post;
 
 public class PostNotFoundException extends Exception {
     public PostNotFoundException(String message) {
