+++ src/main/java/techquack/com/onestar3gram/exceptions/post/UnauthorizedPostException.java
@@ -1,4 +1,4 @@
-package techquack.com.onestar3gram.exceptions;
+package techquack.com.onestar3gram.exceptions.post;
 
 public class UnauthorizedPostException extends Exception{
     public UnauthorizedPostException(String message) {
